Updated test setup to allow for easy sharing of tests between interpreter and JIT
