+++ LMNT/test/test_interpreter.c
@@ -1,0 +1,3 @@
+#include "testsetup_interpreter.h"
+
+#include "test_maths_scalar.h"
