--- LMNT/test/test_maths_scalar.c
@@ -1,281 +1,0 @@
-#include "CUnit/CUnitCI.h"
-#include "lmnt/interpreter.h"
-#include "helpers.h"
-#include <stdio.h>
-#include <stdbool.h>
-
-#define FLOAT_ERROR_MARGIN 0.00001
-
-lmnt_ictx* ctx;
-
-CU_TEST_SETUP()
-{
-    ctx = create_interpreter();
-}
-
-CU_TEST_TEARDOWN()
-{
-    delete_interpreter(ctx);
-    ctx = NULL;
-}
-
-static void test_addss(void)
-{
-    archive a = create_archive_array("test", 2, 1, 3, 1, 0,
-        LMNT_OP_BYTES(LMNT_OP_ADDSS, 0x00, 0x01, 0x02)
-    );
-    const lmnt_def* def;
-    TEST_LOAD_ARCHIVE(ctx, "test", def, a);
-    delete_archive_array(a);
-
-    lmnt_value rvals[1];
-    const size_t rvals_count = sizeof(rvals)/sizeof(lmnt_value);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 1.0f, 2.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 3.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 1000.0f, -2750.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], -1750.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, -1210.0f, 2000.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 790.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, -30.7f, -50.5f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], -81.2, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 0.0f, 0.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 0.0, FLOAT_ERROR_MARGIN);
-
-    // NaN
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, nanf(""), -2.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 1.0f, nanf(""));
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, nanf(""), nanf(""));
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-}
-
-static void test_subss(void)
-{
-    archive a = create_archive_array("test", 2, 1, 3, 1, 0,
-        LMNT_OP_BYTES(LMNT_OP_SUBSS, 0x00, 0x01, 0x02)
-    );
-    const lmnt_def* def;
-    TEST_LOAD_ARCHIVE(ctx, "test", def, a);
-    delete_archive_array(a);
-
-    lmnt_value rvals[1];
-    const size_t rvals_count = sizeof(rvals)/sizeof(lmnt_value);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 1.0f, 2.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], -1.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 300.0f, -2250.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 2550.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, -1210.0f, 2000.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], -3210.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, -3.5f, -5.7f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 2.2, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 0.0f, 0.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 0.0, FLOAT_ERROR_MARGIN);
-
-    // NaN
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, nanf(""), -2.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 1.0f, nanf(""));
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, nanf(""), nanf(""));
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-}
-
-static void test_mulss(void)
-{
-    archive a = create_archive_array("test", 2, 1, 3, 1, 0,
-        LMNT_OP_BYTES(LMNT_OP_MULSS, 0x00, 0x01, 0x02)
-    );
-    const lmnt_def* def;
-    TEST_LOAD_ARCHIVE(ctx, "test", def, a);
-    delete_archive_array(a);
-
-    lmnt_value rvals[1];
-    const size_t rvals_count = sizeof(rvals)/sizeof(lmnt_value);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 4.0f, 2.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 8.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 300.0f, -2.5f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], -750.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, -1210.0f, 2000.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], -2420000.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, -3.5f, -5.7f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 19.95, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 5.0f, 1.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 5.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 5.0f, 0.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 0.0, FLOAT_ERROR_MARGIN);
-
-    // NaN
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, nanf(""), -2.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 1.0f, nanf(""));
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, nanf(""), nanf(""));
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-}
-
-
-static void test_divss(void)
-{
-    archive a = create_archive_array("test", 2, 1, 3, 1, 0,
-        LMNT_OP_BYTES(LMNT_OP_DIVSS, 0x00, 0x01, 0x02)
-    );
-    const lmnt_def* def;
-    TEST_LOAD_ARCHIVE(ctx, "test", def, a);
-    delete_archive_array(a);
-
-    lmnt_value rvals[1];
-    const size_t rvals_count = sizeof(rvals)/sizeof(lmnt_value);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 4.0f, 2.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 2.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 300.0f, -2.5f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], -120.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, -1210.0f, 2000.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], -0.605, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, -3.5f, -5.7f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 0.6140350877192983, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 5.0f, 1.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 5.0, FLOAT_ERROR_MARGIN);
-
-    // Div-by-zero
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 5.0f, 0.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isinf(rvals[0]));
-
-    // NaN
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, nanf(""), -2.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 1.0f, nanf(""));
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, nanf(""), nanf(""));
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-}
-
-static void test_modss(void)
-{
-    archive a = create_archive_array("test", 2, 1, 3, 1, 0,
-        LMNT_OP_BYTES(LMNT_OP_MODSS, 0x00, 0x01, 0x02)
-    );
-    const lmnt_def* def;
-    TEST_LOAD_ARCHIVE(ctx, "test", def, a);
-    delete_archive_array(a);
-
-    lmnt_value rvals[1];
-    const size_t rvals_count = sizeof(rvals)/sizeof(lmnt_value);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 3.0f, 2.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 1.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 300.0f, 100.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 0.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, -5.1f, 3.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], -2.1, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, -5.1f, -3.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], -2.1, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 1.0f, -3.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 1.0, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 2.5f, 1.4f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 1.1, FLOAT_ERROR_MARGIN);
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 0.0f, 3.5f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_DOUBLE_EQUAL(rvals[0], 0.0, FLOAT_ERROR_MARGIN);
-
-    // NaN
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 3.0f, 0.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, nanf(""), -2.0f);
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, 1.0f, nanf(""));
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-
-    TEST_UPDATE_ARGS(ctx, def, (lmnt_offset)0, nanf(""), nanf(""));
-    CU_ASSERT_EQUAL(lmnt_execute(ctx, def, rvals, (lmnt_offset)rvals_count), rvals_count);
-    CU_ASSERT_TRUE(isnan(rvals[0]));
-}
-
-
-CUNIT_CI_RUN("maths",
-    CUNIT_CI_TEST(test_addss),
-    CUNIT_CI_TEST(test_subss),
-    CUNIT_CI_TEST(test_mulss),
-    CUNIT_CI_TEST(test_divss),
-    CUNIT_CI_TEST(test_modss)
-);
